requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
faker==24.4.0
psycopg2-binary
python-dotenv
